refactor(web): lift shared initials() helper to lib/utils (#317)
Address code-review: the new admin settings sidebar duplicated the
initials() avatar helper from app-sidebar.tsx. Hoist it to lib/utils
and import in both, so a future avatar tweak is a one-file change.
